# services/gateway/cmd/main.go
package main

import (
	"context"
	"gateway/internal/common"
	"gateway/internal/modules/auth"
	"gateway/internal/modules/message"
	"gateway/internal/modules/media"
	"gateway/internal/modules/user"
	"gateway/internal/ws"
	"log"
	"net"
	"net/http"
	"os"
	"runtime"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/lxzan/gws"
	"github.com/nats-io/nats.go"
	"github.com/prometheus/client_golang/prometheus/promhttp"
)

func main() {
	natsURL := os.Getenv("NATS_URL")
	if natsURL == "" {
		natsURL = nats.DefaultURL
	}
	var nc *nats.Conn
	var err error
	for i := 0; i < 10; i++ {
		nc, err = nats.Connect(natsURL,
			nats.NoEcho(),
		)
		if err == nil {
			break
		}
		log.Printf("Tentative de connexion à NATS (%d/10) échouée: %v. Nouvel essai dans 5s...", i+1, err)
		time.Sleep(5 * time.Second)
	}
	if err != nil {
		log.Fatalf("Impossible de se connecter à NATS après plusieurs tentatives: %v", err)
	}
	defer nc.Close()
	log.Printf("Connecté à NATS sur %s", natsURL)

	r := SetupServer(nc)

	addr := ":8080"

	// SO_REUSEPORT : N listeners sur le même port = N queues d'accept parallèles
	// Permet d'exploiter tous les CPUs pour accepter les connexions WS simultanées
	const SO_REUSEPORT = 0xf // Linux constant
	lc := net.ListenConfig{
		Control: func(network, address string, c syscall.RawConn) error {
			return c.Control(func(fd uintptr) {
				if err := syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, SO_REUSEPORT, 1); err != nil {
					log.Printf("SO_REUSEPORT non supporté: %v", err)
				}
			})
		},
	}

	numListeners := runtime.NumCPU()
	server := &http.Server{Handler: r}
	errCh := make(chan error, numListeners)

	for i := 0; i < numListeners; i++ {
		ln, err := lc.Listen(context.Background(), "tcp", addr)
		if err != nil {
			log.Fatalf("listen[%d]: %v", i, err)
		}
		go func() { errCh <- server.Serve(ln) }()
	}

	log.Printf("Serveur démarré sur http://localhost%s (%d listeners SO_REUSEPORT)", addr, numListeners)
	log.Fatal(<-errCh)
}

func SetupServer(nc common.NatsConn) *chi.Mux {
	hub := ws.NewHub()
	if err := hub.StartNatsSubscription(nc); err != nil {
		log.Printf("Avertissement: Impossible de démarrer l'abonnement NATS : %v", err)
	}

	handler := ws.NewHandler(hub, nc)
	upgrader := gws.NewUpgrader(handler, nil)

	authHandler := auth.NewHandler(nc)
	userHandler := user.NewHandler(nc)
	messageHandler := message.NewHandler(nc)
	mediaHandler := media.NewHandler(nc)

	r := chi.NewRouter()
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)

	// Auth Routes
	r.Post("/auth/register", authHandler.Register)
	r.Post("/auth/login", authHandler.Login)
	r.Post("/auth/refresh", authHandler.Refresh)
	r.Post("/auth/logout", authHandler.Logout)

	// User Routes
	r.Get("/users/search", userHandler.Search)
	r.Get("/users/{id}", userHandler.Get)
	r.Put("/users/{id}", userHandler.Update)

	// Message (proxy vers message-service)
	r.Post("/api/messages", messageHandler.Send)

	// Media upload (gateway -> NATS -> media-service)
	r.Post("/media/upload", mediaHandler.Upload)

	r.Get("/api/messages/{id}", messageHandler.GetById)
	r.Get("/api/messages", messageHandler.GetByGroupId)

	r.Put("/api/messages/{id}", messageHandler.Update)
	r.Patch("/api/messages/{id}", messageHandler.Update)

	r.Delete("/api/messages/{id}", messageHandler.Delete)
	r.Post("/api/messages/{id}/receipt", messageHandler.AckReceipt)

	// Groups/Conversations (proxy vers message-service)
	r.Post("/api/groups", messageHandler.CreateGroup)
	r.Get("/api/groups", messageHandler.ListGroups)
	r.Get("/api/groups/{id}", messageHandler.GetGroup)
	r.Delete("/api/groups/{id}", messageHandler.DeleteGroup)
	r.Post("/api/groups/{id}/leave", messageHandler.LeaveGroup)

	r.Post("/api/groups/{id}/members", messageHandler.AddGroupMember)
	r.Get("/api/groups/{id}/members", messageHandler.ListGroupMembers)
	r.Patch("/api/groups/{id}/members/{user_id}/role", messageHandler.UpdateGroupMemberRole)
	r.Delete("/api/groups/{id}/members/{user_id}", messageHandler.RemoveGroupMember)

	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte("OK"))
	})

	r.Get("/ws", func(w http.ResponseWriter, r *http.Request) {
		// 1. Extract Token
		token := r.URL.Query().Get("token")
		if token == "" {
			token = r.Header.Get("Authorization")
			if len(token) > 7 && token[:7] == "Bearer " {
				token = token[7:]
			}
		}

		if token == "" {
			http.Error(w, "Unauthorized: missing token", http.StatusUnauthorized)
			return
		}

		// 2. Validate Token locally (no NATS round-trip)
		valResult, err := auth.ValidateToken(token)
		if err != nil {
			log.Printf("Validation JWT Error: %v", err)
			http.Error(w, "Authentication service unavailable", http.StatusServiceUnavailable)
			return
		}

		if !valResult.IsValid {
			http.Error(w, "Unauthorized: invalid token", http.StatusUnauthorized)
			return
		}

		// 3. Upgrade et stockage en session
		socket, err := upgrader.Upgrade(w, r)
		if err != nil {
			log.Printf("Erreur upgrade: %v", err)
			return
		}

		// 4. Store user info in socket session
		socket.Session().Store("userId", valResult.User.ID)
		socket.Session().Store("username", valResult.User.Username)

		go socket.ReadLoop()
	})

	// Metrics Prometheus
	r.Handle("/metrics", promhttp.Handler())
	return r
}